Local Debugging C#
Add support for local CloudScript debugging to the C# SDK.

Details

Amend getReqeustActions in make.js to emit an if block that checks
to see if there is URI specified in a settinsg file for the
ExecuteFunction API. If there is such a setting, then that URI is called
via the DebugDoPost method. Otherwise the else block is executed.

In order to facilite the else clause above, amend API.cs.ejs to declare
the httpResult variable at the top of the function and then set it in
a block. For ExecuteFunction this will be an else block. For all other
APIs this block will just be an line block with no semantic effect.

Add a GetLocalSettingsFileProperty to PlayFabUtils.cs.ejs

Amend PlayFabHttp.cs to have a DebugDoPost API that takes a full URI
and then calls _DoPost. _DoPost has the main guts of the previous DoPost
method which now constructs a full uri and then calls _DoPost.